Adicionador métodos para buscar o carrinho de um usuário pelo Id dele, para atualizar o carrinho e deletar.

+++ src/main/java/com/example/ecommerce1/dto/ProductRequest.java
@@ -6,7 +6,7 @@ public record ProductRequest(
         Long categoriaId,
         String nome,
         String descricao,
-        BigDecimal preco,
+        Double preco,
         int quantidadeEstoque,
         boolean ativo
 ) {
